+++ src/Amphibian.Patrol.Training.Api/Repositories/UserRepository.cs
@@ -5,8 +5,7 @@ using System.Threading.Tasks;
 
 using System.Data;
 using Dapper;
-using Dapper.Contrib;
-using Dapper.Contrib.Extensions;
+using Dommel;
 
 using Amphibian.Patrol.Training.Api.Models;
 
@@ -23,7 +22,7 @@ namespace Amphibian.Patrol.Training.Api.Repositories
 
         public async Task InsertUser(User user)
         {
-            user.Id = await _connection.InsertAsync(user).ConfigureAwait(false);
+            user.Id = (int)await _connection.InsertAsync(user).ConfigureAwait(false);
         }
 
         public async Task UpdateUser(User user)
